Add file:// support and provide a nerfed library
/lib/metal-lib.sh and /lib/metal-md-lib.sh now exist, the former is
safer to include and provides simple functions that all metal-dracut
modules may use. The latter is for md-operations (disk and RAID ops).
This library is more strict and will die if the environment is not
suitable.

File support is also added, meaning
metal.server=file://mydir?label=MYUSB will result in loading objects off
that FSLabel from that directory.

CASMINST-2891 always say yes to wiping

+++ 90metalmdsquash/metal-update-fstab.sh
@@ -6,7 +6,8 @@ fstab_metal_new=/etc/$fstab_metal
 fstab_metal_old=/sysroot/etc/$fstab_metal
 
 if [ -f "$fstab_metal_old" ]; then
-    mount -a -T "$fstab_metal_old"
+    mkdir -pv "$(mount -a -f -v -T /etc/fstab.metal | awk '{print $1}' | tr -s '\n' ' ')"
+    mount -a -v -T "$fstab_metal_old"
 else
-    cp -v "$fstab_metal_new" "$fstab_metal_old"
+    [ -f "$fstab_metal_new" ] && cp -v "$fstab_metal_new" "$fstab_metal_old"
 fi
